refactor: split armor pip layout processing
